# api/openapi.yaml
openapi: 3.0.0
info:
  title: Enterprise ATS Resume Analyzer API
  version: 1.0.0
  description: Production-grade REST API specifications for resume upload, job description parsing, matching analytics, and AI rewriting features.
servers:
  - url: https://api.atschecker.com/v1
    description: Production Server
  - url: http://localhost:8000/v1
    description: Local Dev Server
paths:
...
  /auth/login:
    post:
      summary: Authenticate user and issue access tokens
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
        '401':
          description: Invalid credentials

  # ---------------------------------------------------------------------------
  # RESUME MANAGEMENT
  # ---------------------------------------------------------------------------
...
  /rewriter/bullet:
    post:
      summary: Quantify and optimize a single experience or project bullet point
      tags:
        - Optimization Engine
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - original_text
                - target_role
              properties:
                original_text:
                  type: string
                  example: Developed APIs for internal applications.
                target_role:
                  type: string
                  example: Senior Backend Engineer
                target_company:
                  type: string
                  example: Amazon
      responses:
        '200':
          description: Optimization output
          content:
            application/json:
              schema:
                type: object
                properties:
                  weakness_score:
                    type: integer
                    minimum: 0
                    maximum: 100
                  original:
                    type: string
                  improved:
                    type: string
                  reasons:
                    type: array
                    items:
                      type: string

  # ---------------------------------------------------------------------------
  # BENCHMARKS
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...